feat: move occlusion spot test
