+++ planning/behavior_velocity_occlusion_spot_module/CMakeLists.txt
@@ -22,8 +22,9 @@ if(BUILD_TESTING)
   )
   target_link_libraries(test_${PROJECT_NAME}
     gtest_main
-    behavior_velocity_planner
+    ${PROJECT_NAME}
   )
+  target_include_directories(test_${PROJECT_NAME} PRIVATE src)
 endif()
 
 ament_auto_package()
